Switched to qualified library names to avoid expensive filesystem lookups.
Part of the library loading process is to first call dlopen() on the name. If that returns a NULL address, then FFI has a custom lookup strategy that walks a list of directories. By using names that dlopen() can find, we can take advantage of the ld.so cache and avoid the expensive FFI custome lookup.