Update Truffle
This version is not available on lafo nexus as usual, due to a bug.
Therefore to build using Maven you will need to `mx maven-install`. That
doesn't install with a -SNAPSHOT suffix. We also don't seem to get the
usual transitive dependencies though, so I have added a Maven dependency
on jline for the moment.