Add TruffleRuby logo, by @Talkdesk
Every day at @Talkdesk we use Ruby for many of our production
applications. We strongly believe that TruffleRuby is about to
completely change the Ruby game, and we want to be at the forefront of
the revolution!

We've submitted our proposal for a logo in #51, and we thank everyone
for the feedback there, on twitter, and over other channels 🙇. As a
result of that feedback, we're also including an icon version that drops
the stars background, which is much more readable when using smaller
sizes. As for the base design, we think enough people were on board that
we're really confident that this combination is the way to go! 😃

In this PR, we include five variants of the logo:

* Full logo with TruffleRuby text below rocket
* Full logo with TruffleRuby text beside rocket
* Full logo without text
* Simpler logo without stars background, usable for small sizes
* Sticker version

This entire contribution is licensed under a
Creative Commons Attribution 4.0 International License.