Add a patch for the unf gem.
The unf gem relies on a C extension to provide the String#unicode_normalize method in older versions of Ruby. They haven't released a version of the gem that will use the stdlib version of this method, although it's available in their master repository. By backporting that change here, we can run the library without needing to support the extension.