Avoid creating a substring rope for parsing magic comments.
Previously we were creating a substring rope and then immediately reifying its byte array. The lexer helper methods and the joni methods all work with a byte array and a supplied offset and length, so rather than embed those values into a new substring rope, we can just use the base rope with the supplied offset and length. This saves a new object creation and potentially a byte array copy.