Set the socktype for getaddrinfo to allow it to work on Solaris.
getaddrinfo on Solaris doesn't seem to work without having the socketype set if the service value is a port number. With an empty socktype, the call always errors out with EAI_SERVICE. I experimented with various combinations of values in the addrinfo structure from C and couldn't get a socktype of 0 to work, in any event. Looking at the MRI source for the equivalent call to pack_sockaddr_in, it always set the socktype to SOCK_DGRAM if the current value is 0 and the service name can be parsed as a numeric port value. Doing the same allows the socket code to work fine on Solaris.