Prepopulate the String shape with common hidden properties.
During bootstrap we'll encounter Strings that are frozen and/or tainted as well as Strings that have their object_id calculated. Prior to this change we started with a base shape and transitioned it as each new property was encountered, meaning live Strings were apt to take on a small set of shapes. Currently our dispatch logic caches on the shape, so we could end up with multiple entries in the dispatch chain for each shape type seen. By prepopulating the properties in the shape, we ensure that all Strings we create have the same shape. Note that code elsewhere could still modify the shape after the String is created, in which case that new shape will get its own entry in the dispatch chain as before.