Remove RopeBuffer.
It was only implemented to compare Rope performance against C-style
strings. We haven't run any benchmarks with it in close to a year. At
this point, the code is just getting in the way and could probably be
better implemented within the Rope hierarchy now that we also need to
support Ropes backed by native memory.