`rb_str_cat' should always retain the encoding of the source string.
In MRI, this function simply resizes the underlying string buffer, if necessary, and appends the new bytes to the buffer. This approach avoids any of the Ruby-level steps of encoding negotiation. Since we implement the function with Ruby-level semantics, we need to ensure the return value's encoding is the same as the source string's.