Ugly hack for loading libm on Debian-based systems.
On Debian-based systems, libm.so is a GNU ld script, not a shared library, and thus fails to `dlopen`. We really need to specify the proper SO version to dlopen. Absent a good general purpose solution, this hack fixes the immediate problem with loading libm, which is needed by Nokogiri.