Add a new form to MakeStringNode to build from RopeBuilder using its byte array directly.
In many cases the RopeBuilder is a local object that doesn't escape and makes a defensive copy of any of the data used to populate the builder. So, when we convert to a Rope, there's no need to make an extra copy because the underlying array can't possibly be modified by anything else.