Fix patching of the thread_safe gem.
The old approach allowed the gem to boot up, but ultimately didn't work because the patch for Rubinius relied on Rubinius behavior we didn't support (principally, overriding the `allocate' method). Now that we have a real patching framework, the old constant replacer code can be replaced with something more robust.