Remove unnecessary ConcurrentHashMap for Encoding->Charset.
This was in place because the Encoding->Charset lookup in EncodingManager was a bit expensive. Since that code has been cleaned up, this map isn't really necessary and just complicates the code. It should be noted that Encoding mantains an internal reference, lazily populated, to its calculated Charset. There's an inherent race in calculating the Charset, but the calculated Charset would always be the same reference so running without a lock isn't problematic.