Unroll loop for Array#each where the array only contains a single element.
It's a common pattern to have a method accept varargs and iterate over them. In many cases, these methods are called with a single element that is wrapped into an array. Manually unrolling for this case can yield significant savings. The generated code for a no-op iteration is reduced by 2/3.