Update IO::InternalBuffer to fill its buffer directly from a pointer, rather than an intermediary string.
Previously we were quite wasteful. We'd allocate a new byte[], copy the pointer contents to that byte[], and then construct a rope from that byte[]. Constructing the rope would entail walking the full byte[] to determine the code range. Once the rope was constructed, we allocated a Ruby string to contain the rope. Ultimately this made its way back to IO::InternalBuffer#fill, which subsequently deconstructed the string to get at the underlying rope's byte[], discarding the rope and string in the process and never using the calculated code range. Now we copy directly from the pointer's native memory to the buffer without any wasteful string allocations.