Don't create ConcatRopes when appending a non-empty rope with an empty rope.
We previously had specializations like this and backed them out because they complicated the resulting machine code, causing operations to slow down in some cases. A lot of the supporting code has been improved where this is hopefully no longer an issue. The primary benefit to these specializations is when a non-empty LeafRope is being concatenated with an empty one -- by not creating a ConcatRope we can directly access the LeafRope's byte[] without having to copy it to the root of the tree.