Add a new method to simplify getting a character's length.
Owing to some heritage from MRI, we have multiple methods for getting
a character length, each with slightly different semantics about what
to do with the return value. Moreover, none of them take existing
context into account and always end up validating the bytes being
scanned.

Since we track the code range for ropes, we should be able to use
that context to simplify the task of finding character lengths, so
long as we assume that the byteOffset provided always corresponds
to the first byte in a character and byteEnd is always wide enough
for that character. The caller should have no problem adhering to
that restriction if the rope is either CR_7BIT or CR_VALID. In
particular, this means we can determine MBC UTF-8 character width
simply by looking at the first byte.