Add a dummy class to improve compatibility between JRuby-style and Truffle-style interop.
Specifically, test-unit looks for `Java::JavaLang::Throwable` if it detects that `::Java` is defined. We need `::Java` to be defined for Truffle-style interop, which overlaps with JRuby-style interop. However, the rest of the JRuby-style classes aren't loaded until `require 'java'` is executed, which produces the conflict in test-unit. This isn't an ideal fix, but it allows more test suites to execute properly.