Remove explicit cloning of exception messages from core files.
Semantically, these messages shouldn't be frozen. In practice, it's very unlikely someone will try to mutate an exception's message. In order to clean up the source code, the explicit `+` calls are removed. If this turns out to be problematic, we can look at implicitly adding them back in when building the AST.