Remove the `singleByteOptimizable` field from `Rope`.
The field was added early on because we had it for our `ByteList`-backed strings. As it turns out, it's not used in a ton of places and a sizable percentage of the places it's used are behind a boundary. There are still some paths where it may benefit to be optimized, but that can be handled in a node. Removing it means one less field allocated in every rope, simple rope creation because the value doesn't need to be propagated, and not having to worry about all the places to invalidate now that `NativeRope` has mutable metadata.