Generate Ruby stubs for core methods and primitives
It can be navigated to a definition of a core method. It will open the
stub which has a small javadoc snippet inside which can be used to
navigate to the Truffle node defining the core Ruby method. Same works
for primitives. Idea indexes the stubs therefore auto-completion for
core methods and primitives works.

The primitive encoding was changed from:
    Truffle.primitive :a_name
    Truffle.invoke_primitive :a_name, *args
to:
    TrufflePrimitive.a_name
    TruffleInvokedPrimitive.a_name *args

The stubs are generated into
mxbuild/org.truffleruby/src_gen/core_class_stubs