Add dummy implementation of Module#const_source_location
We don’t know how to actually implement this yet, so we’re just
returning `nil`.

For the sake of simplicity, we are casting the `name` argument to a Java
`String`. This means we don’t currently have access to the original
`RubySymbol` object if the method was called with a `name` that’s a
symbol. In future we may need to undo this cast if we discover that the
original `RubySymbol` is useful, e.g. for its object identity.

Co-authored-by: Maple Ong <maple.ong@shopify.com>