Tag newly-failing Module#const_source_location specs
These tests were accidentally passing before, because they are all
expecting a `NameError`, which the previous `NoMethodError` happens to
be a subclass of. Now that we’ve introduced a dummy implementation of
the method, these tests have begun to correctly fail, since we haven’t
actually written any code to intentionally raise a `NameError`.

Co-authored-by: Maple Ong <maple.ong@shopify.com>