Implement Range#to_a for RubyLongRange
We need a separate specialization for this because the Java object is an
instance of `RubyLongRange` (not `RubyIntRange`) if either of the range
endpoints is too large to fit in a Java `int`. We have basically copied
the implementation of the existing `toA(RubyIntRange)` method because we
don’t think there’s a neat way to reuse this code, since it would need
to be generic over primitive types (`int` vs `long`).

We’re using `Math.toIntExact` to convert the `long` range length into an
`int` because casting it would silently truncate it to 32 bits.
`Math.toIntExact` will instead raise an `ArithmeticException`, which
currently results in a (confusing) `ZeroDivisionError` Ruby exception.
We don’t care very much about this behaviour since it’s not the bug
we’re trying to fix, but will clean this up a little in the next commit.

Co-authored-by: Lillian Zhang <lillian.zhang@shopify.com>