Raise RangeError during Range#to_a if range is too long
Our use of `Math.toIntExact` to convert the `long` range length to an
`int` will raise an Java `ArithmeticException` if the length is too
large to do this without truncation. If the `ArithmeticException` is
allowed to escape, this results in a Ruby `ZeroDivisionError`:

  >> (0..(1 << 40)).to_a
  ZeroDivisionError (divided by 0)

We don’t understand why a `ZeroDivisionError` happens specifically, but
it’s confusing regardless.

By catching the `ArithmeticException` and re-raising it as a Ruby
`RangeError` instead, we can give a more useful error:

  >> (0..(1 << 40)).to_a
  RangeError (long too big to convert into `int')

We copied the error message from the `range_normalized_start_length`
primitive in the same file since it’s dealing with the same problem.

Co-authored-by: Lillian Zhang <lillian.zhang@shopify.com>