Find constant source location with ModuleOperations.lookupConstantWithInherit
As well as supporting inheritance, this checks that the string is a
valid (non-scoped) constant name, which makes several more tests pass.
Since e487b68ccadab953eedbd8ef71b9b5dd883a1126 it’s important that we
pass `true` as the optional `checkName` argument to ensure we get the
latter behaviour.