Refactor random thread-safety to be enforced by RubyRandomizer
The addition of `threadSafe` in the previous commit added a bit of
complexity to both `Randomizer` and `RubyRandomizer`:

* we used `threadSafe` inside the `Randomizer` to toggle thread-safety,
  whenever needed
* we used `threadSafe` inside the `RubyRandomizer` to keep the desired
  state of the `threadSafe`, so that whenever new `Randomizer` instances
  were created (e.g. when setting a new seed), this state was passed
  along.

To avoid this complexity and duplication, let's refactor the
`threadSafe` behavior completely into `RubyRandomizer`:

* `Randomizer` instances no longer need to care about synchronization
  or truffle stuff -- they are now 100% dedicated to the concern of
  keeping the state of the random generation, and outputting new values

* `RubyRandomizer` now takes care of applying synchronization whenever
  needed. All uses of the `Randomizer` object now go through it, and
  this still allows new instances of the `Randomizer` to be set.

I had also looked into the alternative that @eregon suggested -- making
`Randomizer` re-initializeable with a new seed at any time (thus we
wouldn't need to create new instances, and thus not need to track
`threadSafe` inside `RubyRandomizer`), but wasn't entirely happy with
the result, since it made `Randomizer` and the code calling it in
`RandomizerNodes` more complex.

I've also validated that the performance improvements from the
previous commit were preserved.