Combine now-identical Nil and RubyProc specializations
These pairs of specializations used to differ in some way (e.g.
converting the `NotProvided block` into a `nil` or `null` for some
downstream call) but they’re the same now that `nil` is always used to
represent a missing block, so we can combine them and avoid the need to
partially evaluate both paths.