Call divmod when coercion to Float fails for sleep
`sleep': can't convert ActiveSupport::Duration into Float

Is the TypeError TruffleRuby gives when the follow is run:
```
require "active_support/core_ext/numeric/time"
p sleep(0.2.second)
```
In contrast, this prints 0 in MRI 2.6.5 and 2.7.1 (though weirdly prints 1 in 3.0.0). This is because when the argument to sleep is not a fixnum, float, or bignum, it will call .divmod(1) (https://github.com/ruby/ruby/blob/master/time.c#L2556-L2627).

This PR calls divmod instead of to_f (and thus NumericToFloatNode  can be removed).