Add barebone Ruby implementation for ConcurrentHashMap
Following the API in concurrent-ruby's synchronized_map_backend
file. This is the first step to improve the performance of
Concurrent::Map for TruffleRuby. Currently TruffleRuby still uses
the default backend on `concurrent-ruby`. We are creating a backend
on the TruffleRuby side so we are able to optimise the mapping ourselves.

This version of `TruffleRuby::ConcurrentHashMap` passes most specs on
`concurrent-ruby`'s `Concurrent::Map` class except: #dup/#clone,
updates dont block reads, and atomicity.