Make Spin environment file non-executable
This file is intended to be read with `source`, not executed directly,
so it’s misleading for it to have execute permission and a shebang line.

In particular the `SCRIPT_PATH` is designed to work on both Bash and
Zsh, which looks a little odd in a script with a Bash shebang.