Move pattern fittting guard in string_character_index to ruby string#index
The primitive string_character_index is only called in string#index. We can consider offset during the early return, which is actually better than considering it in Java, where we are working in bytespace rather than characterspace. Thus no non-fitting patterns will be checked in the primitive. We can remove the guard and add an assertion in case of future code changes.