Print out the match string's encoding in instrumented output.
Also, fix `RegexpCacheKey` to use the encoding of the regexp rather than the string being matched against. By storing the encoding of the string being matched against in `MatchInfo`, we'll have separate entries for matches that share the same regexp instance, but different match string encodings. That's ultimately what we want in order to gain better insights into how regexps are being used in the application.