Remove unnecessary work in negotiating the encoding to use in a `Regexp` match.
The old approach did not align with MRI and ended negotiating the encoding twice. The result of the first negotiation was used only to detect errors and then subsequently ignored. Moreover, that additional negotiation was semantically different than what MRI does. Part of this clean-up involved implementing MRI's `rb_reg_prepare_enc` and using that for all match operations.

As an additional benefit, by synchronizing with MRI's implementation, we now handle a warning case that we did not previously, allowing us to untag some MRI tests.