Use TRegex's `execBoolean` on matches that do not need a `MatchData` result.
While we could avoid allocating a `MatchData` object when one is not needed in TruffleRuby, we still had to allocate a result in TRegex, which was summarily ignored. Using this new TRegex method will allow for a more efficient regex match on calls such as `Regexp#match?` and `String#match?`.