Added two more Liquid benchmarks, one to parse all files, and one to render a bibliography
This follows how the existing benchmarks are implemented.
A difference with the version in the FastStart project is the result verification.
Here, it’s done inside the benchmark loop, which may already the work load, and which aspects are dominating the benchmarks.

A quick look at the profiles seems how ever indicate that it’s not much changed.

Signed-off-by: Stefan Marr <git@stefan-marr.de>