Update `StringNodes.UnpackNode` for TruffleString.
Also fix an issue with the string format nodes not working properly with a non-zero source byte position. The logic expected the source end position to be specified, not its length. Since a rope's byte array is always bounded exactly, there was never an issue because the source position was always 0 and its end position was always the byte array length. TruffleString objects can subindex a larger byte array, which triggered the bug.