Collapse ASCII-only and MBC w/ ASCII case mapping option in all case change operations.
Using TruffleString's code point iterator allows us to share code for both ASCII-only and MBC strings when using the ASCII-only case mapping option. The iterator will take care of determining code point boundaries and advancing the byte offset accordingly. In addition to reducing the number of specializations we need, this change allows for ASCII-only case mapping operations on MBC strings to PE. The old implementation with ropes was a port from JRuby and was fairly complex, so it was placed behind a `@TruffleBoundary`, but it can be cleanly implemented with TruffleString nodes now.