Fix `String#split` missing a value in the return array under a very specific circumstance.
When `String#split` is called with `' '` as the separator character and a _limit_ value > 0 on an input string that has trailing whitespace and the split string would have fewer entries than the _limit_ value, then we need to append an empty string value to the return array. Ruby doesn't have these semantics when the separator is any other character or a regex, but we need to match for compatibility reasons.