Support null encoding for rb_enc_interned_str_cstr
This commit adds support for a null pointer to be passed as the encoding
argument to rb_enc_interned_str_cstr.

When we implemented this function originally, we noticed that the CRuby
behavior did not match the header documentation, so we opted not to
support it. However, this has now been fixed in CRuby [1], so we can fix
it here as well.

[1]: https://github.com/ruby/ruby/pull/10169