Remove generating platform-dependent 'if RUBY_PLATFORM.include?' guard in exclude files
We've encountered an issue with adding this guard when a test name contains platform (e.g. 'linux' or 'darwin'). If such a test fails then it's excluded on the detected platform but is still run on the other one. So may not be skipped/omitted explicitly and fail. Example of such test - TestThread#test_thread_native_thread_id_across_fork_on_linux.